Move APIs to network package
