--- app/src/main/java/com/example/hospitalnotifier/ScheduleResponse.kt
@@ -1,9 +1,0 @@
-package com.example.hospitalnotifier
-
-data class ScheduleResponse(
-    val scheduleList: List<Schedule>?
-)
-
-data class Schedule(
-    val meddate: String?
-)
