+++ app/src/main/java/com/example/hospitalnotifier/network/TelegramClient.kt
@@ -4,21 +4,6 @@ import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.scalars.ScalarsConverterFactory
-import retrofit2.http.Field
-import retrofit2.http.FormUrlEncoded
-import retrofit2.http.POST
-import retrofit2.http.Path
-
-interface TelegramApi {
-    @FormUrlEncoded
-    @POST("/bot{token}/sendMessage")
-    suspend fun sendMessage(
-        @Path("token") token: String,
-        @Field("chat_id") chatId: String,
-        @Field("text") text: String,
-        @Field("parse_mode") parseMode: String = "Markdown"
-    ): retrofit2.Response<String>
-}
 
 object TelegramClient {
